+++ x/emissions/keeper/actor_utils/losses.go
@@ -199,6 +199,7 @@ func CloseReputerNonce(
 		return err
 	}
 
+	types.EmitNewReputerLastCommitSetEvent(ctx, topic.Id, blockHeight, &nonce)
 	sdkCtx.Logger().Info(fmt.Sprintf("Closed reputer nonce for topic: %d, nonce: %v", topicId, nonce))
 	return nil
 }